# Make single goroutine encodes more efficient