--- a/options.go
+++ b/options.go
@@ -16,6 +16,7 @@ type options struct {
 	usePAR1Matrix                         bool
 	useCauchy                             bool
 	shardSize                             int
+	perRound                              int
 }
 
 var defaultOptions = options{