… galoisAvx512_amd64_test.go